feat(codegen,recipes): extend emitters for Phase 10/11 wrapper param/return shapes + widen 2 wrapper signatures (11-02b Rule 1)
Integration of the 11-02b surface.toml revealed 4 emitter gaps and 2
wrapper-signature misalignments that blocked `cargo check -p lunaris-py
-p lunaris-ts` against the regenerated snapshots. All are Rule 1 fixes
scoped narrowly; the committed Plan 08-01 snapshot stays byte-identical
for the single-`lunaris`-module revision and only changes when 11-02b
flips the emitter into multi-module mode.

Emitter changes (both emit_py.rs and emit_ts.rs where symmetric):

1. Sync-ref_self arm pre-depythonizes Named / Json / Vec / Option /
   Handle params with explicit owned-type annotations. Load-bearing for
   `MeetingNotesMemory::attendees(Vec<String>)` whose PyO3 / napi-rs
   signature surfaces as `&Bound<'_, PyAny>` / `serde_json::Value` at
   the FFI boundary; without pre-bind the call site can't coerce to the
   concrete Rust arg type.

2. Sync-owned + clone_self arm (emit_py.rs) pre-depythonizes the same
   set. TS mirror already handled this via `emit_ts_owned_bindings`.
   Load-bearing for
   `DocumentKnowledgeBase::filter(value: impl Into<serde_json::Value>)`
   whose `impl Into<T>` generic cannot be resolved without the owned-let
   annotation.

3. Async-owned + clone_self arm (emit_ts.rs) now branches on return
   type: `RefSelf` / `Named` land a wrapper-class Arc; other shapes
   (`Vec<Hit>`, etc.) fall through to `ts_async_return_ty_rust` +
   `ts_return_expr_rust` JSON pass-through. Previous code always wrapped
   `out` into the SOURCE type's class, which is a bug for
   `DocumentKnowledgeBase::search(self) -> Vec<Hit>` (wrapping a
   `Vec<Hit>` into `DocumentKnowledgeBase` fails to typecheck).

4. `ts_return_expr_rust` now handles `IrTyRef::Vec` by mapping each
   element through `serde_json::to_value` rather than wrapping the
   whole Vec — matches `ts_async_return_ty_rust`'s `Vec<serde_json::Value>`
   signature. Load-bearing for every async `-> Vec<Hit>` return.

5. `ts_sync_return_expr_rust` handles `IrTyRef::RefSelf` by constructing
   `Self { inner: Arc::new(out) }` (matching the PyO3 equivalent in
   `py_sync_return_expr`). Load-bearing for
   `EmailThreading::thread(&self, root_id) -> Self` where the wrapper
   doesn't derive Serialize.

Wrapper signature widens (≤3 LOC each, pre-authorised by 11-02b PLAN §
risk-mitigations item 3 and the planner's D4(b) refinement):

  - crates/lunaris-recipes/src/conversational/meeting_notes_memory.rs:
    MeetingNotesMemory::attendees — `&[&str]` → `Vec<String>` so the
    codegen `StrList` allow-list entry flows cleanly. Avoids a new IR
    grammar variant for borrowed slice-of-borrow.
  - crates/lunaris-recipes/src/documentary/customer_support_history.rs:
    CustomerSupportHistory::ingest_chat — `turn_idx: u32` → `usize` so
    the codegen `kind = "usize"` annotation matches the Rust arg type.
    Test fixture widened in step with the wrapper.

Test coverage:
- New test `py_and_ts_new_allow_list_entries_lower_to_concrete_rust_spellings`
  (landed in commit ea67088) covers DocumentChunks / UnixMs / StrList /
  JsonValue spellings.
- `ts_async_owned_clone_self` updated to expect
  `napi::Result<Vec<serde_json::Value>>` (the corrected behaviour) for
  an async-owned clone_self method whose return is `Vec<Named>`.

Regression sweep:
- cargo test -p lunaris-codegen — 33 passed.
- cargo check -p lunaris-py -p lunaris-ts — clean.
- cargo run -p lunaris-codegen -- --check — clean on the current
  single-module TOML (non-legacy mode flips on in the follow-up commit
  that adds the two new [[module]] blocks).

author: Tin Dang

+++ crates/lunaris-recipes/src/documentary/customer_support_history.rs
@@ -86,10 +86,15 @@ impl CustomerSupportHistory {
 
     /// Ingest one chat-turn. `ticket_id/turn_idx` slug lands in the
     /// MessageStream thread_id so chats cluster per ticket. (1 primitive call.)
+    // Plan 11-02b Rule 1 widen: `turn_idx` accepts `usize` so the codegen
+    // `kind = "usize"` annotation flows cleanly across the FFI boundary
+    // without a `U32` allow-list entry in the emitter. usize is strictly
+    // wider than u32 on every platform Lunaris supports; production
+    // semantics unchanged (turn indices never overflow 2^32).
     pub async fn ingest_chat(
         &self,
         ticket_id: impl Into<String>,
-        turn_idx: u32,
+        turn_idx: usize,
         participant: impl Into<String>,
         msg: impl Into<String>,
     ) -> Result<Lsn, LunarisError> {
